+++ herbridge/frontend/src/components/App/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import {hot} from 'react-hot-loader'
 import Grid from '@material-ui/core/Grid'
-import PhotoGrid from '../PhotoGrid'
+import PhotoGridList from '../PhotoGridList'
 import TargetResource from '../TargetResource'
 import LogoHerBridge from '../Svg/logo-herbridge.svg';
 import Svg from 'react-svg-inline'
@@ -50,7 +50,7 @@ class App extends React.Component {
               <TargetResource onSearch={this.handleResourceSearch} onResourceSelected={this.handleResourceSelect} onResourceDeselected={this.handleResourceDeselect}/>
             </Grid>
             <Grid item>
-              <PhotoGrid/>
+              <PhotoGridList />
             </Grid>
           </Grid>
         </MuiThemeProvider>
