--- herbridge/frontend/src/components/PhotoGrid/index.js
@@ -1,2 +1,0 @@
-import PhotoGrid from './PhotoGrid';
-export default PhotoGrid;
