+++ herbridge/frontend/src/components/PhotoGridList/index.js
@@ -1,0 +1,2 @@
+import PhotoGridList from './PhotoGridList';
+export default PhotoGridList;
